Update GHA configuration
"This request has been automatically failed because it uses a deprecated
version of `actions/cache: v2`. Please update your workflow to use v3/v4
of actions/cache to avoid interruptions"

I suppose that is the only way of getting the users to wake up and stop
using the deprecated versions.